refactor(#3479): extract TUI-direct anchor completion lifecycle from tui_prompt_relay.rs
Move the live-relay ⏳→✅ anchor completion cluster (should_complete_tui_direct_anchor_lifecycle, DeferredAnchorCompletionDrain + decide_deferred_anchor_completion_drain, complete_tui_direct_prompt_anchor_lifecycle_if_present, pinned_anchor_cleanup_target, complete_tui_direct_anchor_lifecycle_for_inflight) verbatim to tui_prompt_relay/anchor_completion.rs (213 prod LoC). Deps via use super::*; only super::formatting becomes super::super::formatting. The three externally-called helpers are pub(in crate::services::discord), re-exported so the tmux_watcher.rs/recovery_engine.rs call sites stay byte-identical; the relay-internal drain decision/enum re-imported privately. Lowers the tui_prompt_relay.rs giant ratchet 4630 -> 4458 (-172). Zero logic change.
