+++ scripts/audit_maintainability_giant_baseline.toml
@@ -160,7 +160,14 @@
 # `tui_prompt_relay/rehydration.rs` (~296 LoC), both reached via re-import so the
 # call sites stay byte-identical. Lowers the frozen baseline 5142 -> 4630 (-512).
 # Locks in the shrink win (zero logic change).
-"src/services/discord/tui_prompt_relay.rs" = 4630
+# #3479 item-2 (behavior-preserving extraction): the live-relay TUI-direct prompt
+# anchor COMPLETION lifecycle (`⏳ → ✅`) cluster — the visibility gate, the
+# deferred `⏳`-completion drain decision, and the reaction-swap completers —
+# moved verbatim to `tui_prompt_relay/anchor_completion.rs` (213 prod LoC, below
+# the giant threshold), reached via re-import so the call sites stay
+# byte-identical. Lowers the frozen baseline 4630 -> 4458 (-172). Locks in the
+# shrink win (zero logic change).
+"src/services/discord/tui_prompt_relay.rs" = 4458
 # #3038: lowered 4728 -> 4657 (-71) as `SensitivityState` (definition + impl)
 # moved to the `voice_sensitivity` sibling module. Locks in the shrink win.
 # #3038 VoiceBargeInRuntime S1: lowered 4657 -> 4350 (-307) as the STT
